schema tests (sqlite only) and inclusion/exclusion params
